pyproject.toml
[project]
name = "trabalho2ed-minio-sql"
version = "0.1.0"
description = "Projeto de Engenharia de Dados com Spark, Delta Lake, MinIO e SQL Server"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "delta-spark==3.2.0",
    "jupyterlab>=4.0.0",
    "mkdocs-material>=9.7.6",
    "pandas>=2.2.0",
    "pyodbc>=5.1.0",
    "pyspark==3.5.3",
    "python-dotenv>=1.0.0",
    "taskipy>=1.14.0",
]

[tool.taskipy.tasks]
notebook = "jupyter lab notebook/"
jupyter = "jupyter lab notebook/"
docs_build = "python run_mkdocs.py build"
docs_serve = "python run_mkdocs.py serve"
docs_deploy = "python run_mkdocs.py gh-deploy"
